+++ src/onegov.ticket/onegov/ticket/__init__.py
@@ -2,7 +2,13 @@ from onegov.ticket.handler import Handler, HandlerRegistry
 handlers = HandlerRegistry()  # noqa
 
 from onegov.ticket.model import Ticket
-from onegov.ticket.collection import TicketCollection
+from onegov.ticket.collection import TicketCollectionSubset, TicketCollection
 
 
-__all__ = ['Handler', 'handlers', 'Ticket', 'TicketCollection']
+__all__ = [
+    'Handler',
+    'handlers',
+    'TicketCollectionSubset',
+    'Ticket',
+    'TicketCollection'
+]