+++ src/onegov.ticket/onegov/ticket/collection.py
@@ -1,6 +1,9 @@
 import random
 
+from onegov.core.collection import Pagination
 from onegov.ticket.model import Ticket
+from sqlalchemy import desc
+from sqlalchemy.orm import joinedload, undefer
 
 
 class TicketCollection(object):
@@ -89,3 +92,33 @@ class TicketCollection(object):
             query = query.filter(Ticket.handler_code == ensure_handler_code)
 
         return query.first()
+
+
+class TicketCollectionSubset(TicketCollection, Pagination):
+
+    def __init__(self, session, page=0, state='open'):
+        self.session = session
+        self.page = page
+        self.state = state
+
+    def __eq__(self, other):
+        return self.state == other.state and self.page == other.page
+
+    def subset(self):
+        query = super(TicketCollectionSubset, self).query()
+        query = query.order_by(desc(Ticket.created))
+        query = query.options(joinedload(Ticket.user))
+        query = query.options(undefer(Ticket.created))
+        query = query.filter(Ticket.state == self.state)
+
+        return query
+
+    @property
+    def page_index(self):
+        return self.page
+
+    def page_by_index(self, index):
+        return TicketCollectionSubset(self.session, index, self.state)
+
+    def for_state(self, state):
+        return TicketCollectionSubset(self.session, 0, state)