+++ src/onegov.ticket/HISTORY.rst
@@ -4,6 +4,9 @@ Changelog
 Unreleased
 ~~~~~~~~~~
 
+- Adds a ticket collection that supports pagination and filter.
+  [href]
+
 0.0.1 (2015-07-10)
 ~~~~~~~~~~~~~~~~~~~